# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [Users]
      summary: Get a single user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /users/me/role:
    patch:
      operationId: updateMyRole
      tags: [Users]
      summary: Update the current user's role (first-time setup or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [Users]
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Courses ──────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    User:
      type: object
      required: [id, replitId, name, email, role, createdAt]
      properties:
        id:
          type: integer
        replitId:
          type: string
        name:
          type: string
        email:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [student, instructor, admin]
        createdAt:
          type: string
          format: date-time
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [student, instructor, admin]
...