# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  replitId: string;
  name: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  role: UserRole;
  createdAt: Date;
}