# lib/api-zod/src/generated/types/userRoleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Learning Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRoleUpdateRole } from './userRoleUpdateRole';

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}